[WIP] Moved all routes out.
