+++ nlp4all/models/Tweet.py
@@ -1,6 +1,6 @@
 """Tweet model"""
 
-from sqlalchemy import Column, Integer, String, ForeignKey, JSON
+from sqlalchemy import Column, Integer, String, ForeignKey, JSON, DateTime
 from sqlalchemy.orm import relationship
 
 from .database import Base
