+++ requirements-dev.txt
@@ -6,3 +6,5 @@ pylint-flask==0.6
 pylint-flask-sqlalchemy==0.2.0
 pylint-plugin-utils==0.7
 pre-commit==2.20.0
+sqlalchemy-stubs==0.4
+mypy-extensions==0.4.3
